# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "get": {
                "description": "Simple root endpoint that returns a welcome message.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "home"
                ],
                "summary": "Welcome endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WelcomeResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns a basic status payload to indicate the API is running.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "home"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HealthResponse"
                        }
                    }
                }
            }
        },
        "/messages/sent": {
            "get": {
                "description": "Returns a paginated list of successfully sent messages.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "List sent messages",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size (max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SentMessagesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/scheduler": {
            "post": {
                "description": "Starts or stops the background scheduler based on the given action.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduler"
                ],
                "summary": "Control scheduler",
                "parameters": [
                    {
                        "description": "Scheduler action (start|stop)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SchedulerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SchedulerControlResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.SchedulerRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "Action controls the scheduler. Allowed values:\n- \"start\": start processing batches\n- \"stop\":  stop processing batches",
                    "type": "string"
                }
            }
        },
        "response.HealthPayload": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "response.HealthResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.HealthPayload"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "response.MessageDTO": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "messageId": {
                    "type": "string"
                },
                "sentAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "response.SchedulerControlPayload": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.SchedulerControlResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.SchedulerControlPayload"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "response.SentMessagesPayload": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.MessageDTO"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.SentMessagesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.SentMessagesPayload"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "response.WelcomePayload": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.WelcomeResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.WelcomePayload"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    }
}